backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "google-adk>=1.0.0",
    "greenlet>=3.2.2",
    "numpy>=2.2.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pyrr>=0.10.3",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "svg3d>=0.1.2",
    "svgwrite>=1.4.3",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "alembic>=1.16.1",
    "black>=25.1.0",
    "mypy>=1.15.0",
    "ruff>=0.11.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
